@@ -3,12 +3,12 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 30
 
     defaultConfig {
         applicationId "com.example.android.contentproviderpaging"
         minSdkVersion 26
-        targetSdkVersion 28
+        targetSdkVersion 30
     }
 
     buildTypes {
@@ -19,7 +19,7 @@ android {
     }
 
     dependencies {
-        compile project(':resources')
+        implementation project(':resources')
         implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     }
 }
