@@ -16,7 +16,7 @@
 
 package com.example.android.contentproviderpaging
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
