@@ -17,7 +17,7 @@
 package com.example.android.contentproviderpaging
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 
 /**
  * The launcher Activity.
